:tada: Improve timer functions #173
